refactor: review 도메인 api/types 폴더 평탄화
